Simulate a stack of plates
SetOfStacks is the data structure used to simulate a stack of plates.
SetOfStacks should be composed of several stacks and should create a new
stack once the previous one exceeds capacity. Push and Pop for SoS
should behave same as a single stack.

Follow up:
Implment PopAt that performs a pop operation on a specific sub stack